Fixed box rotations!
- Box rotates by some angle about an arbitrary axis
- Fixed box normals when applying transformations

+++ src/expraytrace.cpp
@@ -125,7 +125,7 @@ int main()
 
         raytraceShader.use();
         ShaderSetCamera(raytraceShader, camera);
-        raytraceShader.setInt("Time", int(currentFrame));
+        raytraceShader.setFloat("Time", currentFrame);
         raytraceShader.setVec2("screenSize", screenSize);
         quad.Draw();
